requirements.txt
numpy<=1.20
llvmlite
numba
torch>=1.1
tensorboardX
easydict
pyyaml
scikit-image
tqdm
torchvision
SharedArray
open3d-python
opencv-python
